+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 
 ## Unreleased
 
+## v0.4.0 (September 27, 2024)
+
+* Adds `-n <name>` / `--name <name>` arguments to the CLI to run a specific test
+
 ## v0.3.0 (August 16, 2024)
 
 * Adds platform flags for tests and individual steps